Extract MySQL-specific code into Adapters module
