+++ lib/active_record/database_validations.rb
@@ -1,8 +1,8 @@
 require 'active_record'
 require 'active_support/i18n'
 require 'active_record/database_validations/version'
+require 'active_record/database_validations/adapters'
 
-require 'active_record/database_validations/mysql'
 require 'active_record/validations/database_constraints'
 require 'active_record/validations/string_truncator'
 require 'active_record/validations/typed_column'
